Refactored player spawning so that the player cannot be spawn killed

+++ project3_ds_dj/include/Entitys/Player.h
@@ -41,7 +41,7 @@ private:
 	float m_rotationDiff;
 	const float INPUT_THRESHOLD = 30.0f;
 	const uint16_t FIRE_RATE = 200;
-	const uint8_t TIME_TO_SPAWN = 2;
+	const uint16_t TIME_TO_SPAWN = 2;
 	sf::Clock m_clock;
 	float m_timeToNextShot;
 	BulletType m_bulletType = BulletType::SINGLE_BULLET;
